requirements.txt
python-telegram-bot
peewee
pyzmq
